Day 87/requirements.txt
seaborn==0.10.1
streamlit==0.66.0
pandas==1.0.3
numpy==1.18.3
matplotlib==3.2.1
joblib==0.14.1
Pillow==7.2.0
sklearn==0.23.2
